backend/requirements.txt
fastapi==0.104.1
uvicorn==0.23.2
pydantic==2.4.2
python-dotenv==1.0.0
sqlalchemy==2.0.22
alembic==1.12.0
python-multipart==0.0.6
requests==2.31.0
supabase==1.0.3
